+++ .gitignore
@@ -2,3 +2,5 @@ node_modules
 .env
 json.sqlite
 ./commands/medal/roster.js
+package.json
+package-lock.json
